+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Implements array access for `Log` [#1](https://github.com/stevewest/changelog/issues/1)
 - Added name param to Release constructor.
 
+## Changed
+- Renamed "providers" to "IO"
+
 ## [0.1.0][0.1.0] - 2015-01-18
 ### Added
 - Initial version